requirements.txt
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.34.0,<0.35.0
sqlalchemy[asyncio]>=2.0.0,<2.1.0
asyncpg>=0.30.0,<0.31.0
alembic>=1.14.0,<1.15.0
pydantic>=2.10.0,<2.11.0
pydantic-settings>=2.7.0,<2.8.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[argon2]>=1.7.0,<1.8.0
argon2-cffi>=23.1.0
redis>=5.2.0,<5.3.0
httpx>=0.28.0,<0.29.0
PyNaCl>=1.5.0,<1.6.0
python-multipart>=0.0.9
